feat(core-elements): CheckboxBase에서 CheckboxContext 의존성 제거
